requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4
python-dotenv
python-multipart
